Perl does not seem to like dashes in the names of modules, so we won't be naming the library perfSONAR-PS anymore.
This update also contains the new 'XML' sub library that features two 
modules:
  Handler - A SAX handler, meant to be used with the XML::SAX parser.
  Element - Mimics the organization of the java object of the same name.

It is now possible to create a structure of these generic objects and 
operate on them in a similar manner to that of the java release of 
perfSONAR, the only difference being that it is not as namespace 
specific (although after working with it, this doesn't seem to be a 
problem).  

Depending on feedback, I belive we can use the generic approach and not 
bother with creating namespaced specific objects (see the perl doc for 
Element.pm for an example).

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@61 752d5dfb-d90f-4296-8230-53451a7f88fe