Updates to the RRD collection script (and module).  Because rrdtool is limited in how updates work, statements such as 'rrdtool update some.rrd -t eth0-in 1170954063:1' followed by 'rrdtool update some.rrd -t eth0-out 1170954063:2' won't work properly since time has moved on, and you can't reverse it to make an update on an unrelated ds.
Therefore the module has been changed slightly to keep track of ds 
values as well as a single timestamp as a part of the 'insert' 
procuedure.  After you have gathered all ds values for a timestamp, you 
can then 'insertCommit' all values in one shot.  

This problem is of course natural to how we describe the metadata/data 
relationship between the subjects and the eventual RRD files the data 
will live in.  These changes will eventually make their way into the MA 
module as well.  

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@67 752d5dfb-d90f-4296-8230-53451a7f88fe