Moving some 'MA' labeled functions directly into the Transport module:
1) 'Echo' service belongs in the Transport module because all services 
will benefit from being able to be queried to see if they are still 
alive.

2) Messages from the 'MA General' module will be used by other services 
such as the LS, so they have bene seperated out.

3) The common functions 'extract' was moved to Common (it was duplicated 
in both the MP and MA for some reason.

These changes *may* break some of the other services, I will be 
searching through to make necessary changes.

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@207 752d5dfb-d90f-4296-8230-53451a7f88fe