Changes:
  - Log::Log4perl functionality in each of the base and (for now) some 
of the upper level classes.  Explanation of use will show up on both the 
wiki and the tutorial.

  - MA classes:

    * Lots of refactoring to move functions to the Transport module, and 
allow for other 'MA Like' (i.e. the LS) modules to not re-invent the 
wheel.  

  - MP classes:

    * Some, but not a lot of refactoring to the General and Base 
classes.  The MP still requires a lot of specific code for dealing with 
the measurement it is after.

  - Common classes:

    * Common and Message are unchanged beyond logging additions.

    * Transport takes on more responsibility for handling messages.  The 
'echo server' is now handled here, and messages that do not meet minimum 
requirements (# of message elements, correct namespace) are rejected.  
The 'correct' messages are then passed on to the waiting upper levels 
(SNMP MA, etc).  

Todo:

Ping/OWAMP/BWCTL are still not coverted completely.  I will be fixing 
ping to reflect the tutorial and will contact Warren on BWCTL and OWAMP 
changes.

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@213 752d5dfb-d90f-4296-8230-53451a7f88fe