Support for queries complete.  There are some minor differences between the behavior of the Java LS and the perl LS:
- XMLDB doesn't recognize the nmwg:store element, this can complicate 
queries such as /nmwg:store/nmwg:metadata.  There are three ways we can 
fix this problem:
  - Inject 'nmwg:store' elements around everything in the database
  - Specify in our version of the protocol it won't work to say 
'nmwg:store'
  - Figure out a way with the Java developers to be rid of the concept 
of a store element (probably the hardest one since eXist deals with it 
just fine)

- To do a directional query we can't say /nmwg:metadata/netutil:subject, 
we need to do add "collection('COLLECTION_NAME')" before the first 
slash.  For now we can just insert this into incomming queries since 
this is a limitation of XMLDB and not of the xquery standard in general.

- To do a query such as:

  for $metadata in /nmwg:store/nmwg:metadata
    let $metadata_id := $metadata/@id
    let $data := /nmwg:store/nmwg:data[@metadataIdRef=$metadata_id]
    where $metadata//psservice:serviceType="MA"
    return <lookup-storage-from-query>
      {$metadata} {$data}
      </lookup-storage-from-query>

  The XML elements are not parsed correctly by LibXML.  We can't really 
do much about this, except to prehaps preparse the message and convert 
the XML markup to escape sequences (< = &lt;) or just mandate they are 
written this way in the sent query.

- Output is not escaped as in the Java LS, i.e. we dont receive 
pre-escaped content.  I am not sure why the Java LS does this (probably 
how it comes out of the DB) but our XML comes out of the DB as XML.

I will be testing this version against the java version to produce 
better error messages and iron our the rest of the differences.

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@270 752d5dfb-d90f-4296-8230-53451a7f88fe