Adding support for obfuscated keys.  This complicates some things for the short term, so to use this new feature a message level parameter will be needed.  For example, including this parameter at the message level:
<nmwg:parameter name="obfuscatedKey">true</nmwg:parameter>

Will return a key that looks similar to this (instead of a direct copy 
of the store file):

<nmwg:key xmlns:nmwg="http://ggf.org/ns/nmwg/base/2.0/" id="k1">
  <nmwg:parameters id="pk1">
    <nmwg:parameter name="maKey">
      a63b47372a944d040d41d9fefdab7082
    </nmwg:parameter>
  </nmwg:parameter>
</nmwg:key>

This 'maKey' is calculated at start time by the MA using MD5 
fingerprints and the contents of the store file key.  This key can of 
course be replayed over and over, and can be used in conjunction with 
chaining and will be 'cooked' like other keys.

Note that the old style of keys are still valid, and will be returned by 
default (at least until the various clients are confirmed to work with 
this).

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/branches/merge@952 752d5dfb-d90f-4296-8230-53451a7f88fe