Add the beginnings of eventType equivalence. When a merge is performed, if the two elements being merged aren't equivalent, an error will be thrown. A module can register that two eventTypes are equivalent with:
$handler->registerEventEquivalence("MessageType", "eventType1", "eventType2");

When a merge occurs between two metadata in a message of the specified type, if
two eventTypes have been listed as equivalent, the merge will no longer fail if
two metadata are merged and each one has one of the equivalent eventTypes.

If two metadata are being merged and each has multiple eventTypes, only the
common eventTypes will appear in the merged metadata. Thus, if an eventType
appears in one metadata and in the other metadata, an eventType in the same
equivalence class appears, both eventTypes will appear in the merged metadata.

The registerEventEquivalence parameter MessageType has a special value '*' that
means "this eventType equivalence applies no matter what type of message the
metadata appears in".



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/branches/merge@1049 752d5dfb-d90f-4296-8230-53451a7f88fe