An initial implementation of the DCN Topology Agent. It does not yet read the local data from the IDC (just from a file).
Once it reads the local data, it places the information in the local topology
service if it is newer than that currently in there. It then goes through the
topology and looks for all the neighbor domains. Once it has found the neighbor
domain ids, it looks up in a lookup service for the authoritative topology
service for that domain. It then contacts that TS and downloads the full
contents of the TS and loads each domain it finds into the local TS if the
domain info is newer than the data in the local TS. If each domain runs these
agents repeatedly, the full set of topologies will eventually propogate around.



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/branches/merge@1050 752d5dfb-d90f-4296-8230-53451a7f88fe