Registration fix that prevents duplicates from registering when they differ ONLY by id.  Background: services like the topology service generate 'random' containers for their metadata and register with the LS.
The same exact entity (a link, or a domain) may differ only by metadata
ID (due time time of registration).  To prevent the LS from allowing
these duplicates we clone the data, remove any ID tags, then check to
see if it lives in the database.  We will add it only if it does exist.  
There should be little overhead for this operation, although performance 
testing will reveal this more deeply.

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/branches/merge@1066 752d5dfb-d90f-4296-8230-53451a7f88fe