Updates to the summarized data structure.  Too much to go into in this log, see the documentation for full details.  In general we need to maintain 3 levels of data: 	1 Registered Data (from the services directly) 	2 Summarization of Each regisered service (each service gets a summary, this is true for something like an MA or an hLS) 	3 Summarization of the entire data store (all services should be summarized into one all knowing summary [to rule them all])
I am splitting this between two containers: glsstore.dbxml and 
glsstore-summary.dbxml (keeping the names static from now on, this won't 
hurt anyone).  The first will maintain store files for 'LSStore'and 
'LSStore-control' as usual (handling option 1).  The second will maintin 
the previous two (for option 2) as well as an 'LSStore-summary' (used to 
store the results of option 3).  Note that this is geared for internal 
use currently, although it will be exposed via the level 0 API.

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@2031 752d5dfb-d90f-4296-8230-53451a7f88fe