Preliminary updates to gLS/LS to fix several issues:  - Large number of db collisions due to clean/summarization process potentially running at the same time  - Summarizing 'deleted' data  - Removing/Not removing the appropriate data sets
The fix is proposed to be a unified process to handle both (e.g. a 
'maintenance' method, this may be extended to other services in the 
future). 

Fixed an annoying bug in MA::General that didn't really explain why data 
wasn't being returned.

Fixes to theconfigure/daemon scripts to support gLS/LS changes. 

Tidy/Critic on all files.

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@2216 752d5dfb-d90f-4296-8230-53451a7f88fe