updateData.cgi:     Now uses the client MA API. Also - now keys on ifName instead of     ifIndex. (completely ignores 'host' for now, so that may need     to be fixed if we start querying an MA with more than one     host in there.)
speed.js
    ifName instead of ifIndex

Currently this only for sure works with speed.html. (Utilization graph
has not been modified to use ifName, so it will not work.)

Also - the reason speed.html works is that it does not attempt to
set any of the ifName/resolution/direction things that are queried
for. In other words, the defaults at the top of updateData.cgi are
what matters.




git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@2240 752d5dfb-d90f-4296-8230-53451a7f88fe