Minor Changes: 	doc/perltidyrc - decreasing the tightness on parenthesis from 1 to 0 (will add more spaces when tidy'ed) 	lib/perfSONAR_PS/Client/Echo.pm - Better error messages 	scripts/perfSONAR-daemon.pl - Getting rid of an empty string error
Major changes where made to gLS.pm to re-organize the summarization and 
LS registration procedures.  The overall goal is to minimize the 
critical section of DB access.  The previous method left the DB open for 
a long period of time (in some cases the length of the method call) when 
only a single open/query/close was needed in most cases.  

For the summarization especially this allowed the DB to block on almost 
all other operations while it was open (this is a seperate issue of 
DBXML not handling concurrency well, but our fixes here are still 
necessary from a style and design point of view).  

Additional changes are still required for the cleaning method (needs to 
be condensed to a smaller critical section) and there is also a portion 
of the synch code that will benefit from a different design.  

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@2327 752d5dfb-d90f-4296-8230-53451a7f88fe