gLS.pm: More changes addressing transaction management (removing a call to 'isValid' in place of an inline query).  Removing some training wheels around the maintenance/registration variables to allow them to be turned off when a time of '0' is seen in the conf file.
daemon: tidy cleanups + support for the above gLS changes regarding 
maintenance and registration

Document/RequestHandler: General cleanups.  Support for adding all 
namespaces to the message level (a requirement for compat w/ most EU 
services).  Note that currently the message handling itself (e.g. 
creation of the start/end elements) is done in the specific services, so 
this change to add all namespaces is only valid in the gLS/hLS as it is 
required for API compatibility.  We *should* move this to the genearl 
daemon in time.  

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@2351 752d5dfb-d90f-4296-8230-53451a7f88fe