Updates to cope with some requirements for loading multiple cricket instances.  Summary:
1) Cricket normally relies on an env variable, $ENV{CRICKET_HOME}.  For 
most situations this works fine for the libraries
2) Since we will occasionally have more than one cricket (e.g. SC data) 
available the env variable would change depending on the data home.  It 
would be ideal to do this:

use lib "/path/to/cricket"

Where this path is specified in the daemon.conf and could change for 
each instance.  'use lib' and BEGIN blocks are loaded at compile time 
though and can't really use external variables.  
3) The included fix relies on manual shifting of the @INC (its perl, 
there are more than 10000 ways to solve the problem!) and the use of 
'require' instead of use.  

-jason



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@2414 752d5dfb-d90f-4296-8230-53451a7f88fe