Updates the ls registration daemon package to deal with some issues that matt mathis and rich encountered.
It also moves the daemon, and some of the modules that are only useful to it,
into this package instead of being in the global directory structure. The rest
of the common files are symlinked from their global position. This way, if we
want to delete the ls registration daemon, it won't leave its unused components
in the global space.



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@2425 752d5dfb-d90f-4296-8230-53451a7f88fe