Minor change to the README, clarifying how the LS works (basically, a copy-paste from the INSTALL)
INSTALL:
Change Up2Date -> up2date
Change references to Apt to APT-RPM to make it clear that it's not a debian or ubuntu package.
Make YUM the first described install method since it will presumably be the most common rpm method
Remove the "--logger" argument in the manual starting mode since the logging will default to the screen if that argument is missing, and the subsequent text told them to make the logger print to the screen.



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/branches/RELEASE_3.1@2713 752d5dfb-d90f-4296-8230-53451a7f88fe