Pull some changes from pS-Toolkit-3.1.3 into trunk:
r3951:
Pull in the store file reread option for the SNMP MA

r3952:
Add an option to the SNMP MA to refresh the cacti/cricket store files periodically, similar to what pSB does. This removes the need to restart the SNMP MA whenever cacti is updated, and allows for the SNMP MA to run constantly instead of exiting if no cacti/cricket stuff is available. Note: only tested on cacti.

r3967:
Only update the store file if the newly generated store file is different from the existing one. 

r3968:
Fixes to the store file generator.



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@4018 752d5dfb-d90f-4296-8230-53451a7f88fe