Bring in the PingER changes from the Toolkit into trunk
r4258:
Change the log rotation to daily for the pS services

r4176:
A smattering of minor bugfixes

r4174:
MySQL has priority "64" in the boot order, so bump the services that depend on
it to later in the boot process.

r4145:
RedHat's shutdown scripts require a file in "/var/lock/subsys" before they will
run any init scripts to shut down the service (it just silently doesn't call
them if it's missing).  Update the various init scripts to place that file
there so that the init scripts will actually get called on shutdown.

r4136:
Make the RPM work more like the other RPMs



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@4456 752d5dfb-d90f-4296-8230-53451a7f88fe