Pull in Lookup Service changes from the pS-Performance_Toolkit branch
r4444:
Add a sysconfig element so the database cleaning/removal can be configured for
the init script


r4428: 
Have the lookup service clear the database contents on startup. We saw an
instance in the field where the LS was hanging on startup.


r4258: 
Change the log rotation to daily for the pS services


r4145: 
RedHat's shutdown scripts require a file in "/var/lock/subsys" before they will
run any init scripts to shut down the service (it just silently doesn't call
them if it's missing).  Update the various init scripts to place that file
there so that the init scripts will actually get called on shutdown.



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@4459 752d5dfb-d90f-4296-8230-53451a7f88fe