A user reported an issue where there hosts' NIC was on interface eth1. By default, PingER assumes 'eth0', and for any other interface name, special configuration is needed. This configuration was not needed in the previous version which meant that if a user whose machine used eth1 were to do a yum upgrade, and then restart PingER, their PingER installation would no longer work until they edited the configuration and specified that they were using eth1 and not the default eth0.
This changes PingER to use the output from 'ping' and 'ping6' to figure out the
source address. If the user configures an interface, ping/ping6 will use that
interface, however, if the user does not specify any interface, ping will use
the default route for that address. This should ensure that if a user leaves
the interface option unconfigured, his PingER instance will collect data as
expected.



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@4961 752d5dfb-d90f-4296-8230-53451a7f88fe